Web/PaperGather/requirements.txt
Flask==3.0.0
Flask-Moment==1.0.5
python-dotenv==1.0.0
loguru==0.7.2
psycopg2-binary==2.9.9
redis==5.0.1